anthropic_api/pyproject.toml
[project]
name = "claude-cli"
version = "0.1.0"
description = "Simple CLI tool for interacting with Anthropic's Claude API"
readme = "README.md"
authors = [
    { name = "grdddj", email = "jiri.musil06@seznam.cz" }
]
requires-python = ">=3.8"
dependencies = [
    "anthropic>=0.39.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
claude-cli = "cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
